moving checkbox field into base_fields dir
